Modified RandomCharSequenceGenerator to request a random 32-bit signed integer instead of a 64-bit signed integer.

+++ Src/AutoFixture/RandomCharSequenceGenerator.cs
@@ -37,7 +37,7 @@ namespace Ploeh.AutoFixture
                 return new NoSpecimen(request);
 
             return Convert.ToChar(
-                this.randomPrintableCharNumbers.Create(typeof(long), context));
+                this.randomPrintableCharNumbers.Create(typeof(int), context));
         }
     }
 }
